requirements.txt
django<=2.3
django-bootstrap4
django-admin-sortable2
django-autocomplete-light
pillow
